refactor: extract shared P12 utility and rename signer adapter
  Extract duplicated P12 parsing logic (OID constants, RFC4514 formatting,
  certificate/key extraction) from p12-reader.ts and p12-parser.adapter.ts
  into a shared utility at src/shared/infra/p12.ts to follow DRY principle.

  Rename LocalSignerAdapter to SignerAdapter since Sealify was removed and
  there is only one InvoiceSignerPort implementation.:
