+++ src/modules/certificate/infra/adapters/p12-parser.adapter.ts
@@ -1,62 +1,17 @@
-import * as forge from 'node-forge';
+import { formatRfc4514, parseP12 } from '#/shared/infra/p12';
 
 import { P12Metadata, P12ParserPort } from '../../domain/ports';
 
-const OID_SHORT_NAMES: Record<string, string> = {
-  '2.5.4.3': 'CN',
-  '2.5.4.6': 'C',
-  '2.5.4.7': 'L',
-  '2.5.4.8': 'ST',
-  '2.5.4.10': 'O',
-  '2.5.4.11': 'OU',
-  '1.2.840.113549.1.9.1': 'emailAddress',
-  '2.5.4.5': 'serialNumber',
-};
-
-function formatRfc4514(attributes: forge.pki.CertificateField[]): string {
-  return [...attributes]
-    .reverse()
-    .map((attr) => {
-      const name = attr.shortName || OID_SHORT_NAMES[attr.type as string] || attr.type;
-      return `${name}=${attr.value}`;
-    })
-    .join(',');
-}
-
 export class P12ParserAdapter implements P12ParserPort {
   parse(buffer: Buffer, password: string): P12Metadata {
-    const p12Der = forge.util.decode64(buffer.toString('base64'));
-    const p12Asn1 = forge.asn1.fromDer(p12Der);
-    const p12 = forge.pkcs12.pkcs12FromAsn1(p12Asn1, password);
-
-    const certBags = p12.getBags({ bagType: forge.pki.oids.certBag });
-    const keyBags = p12.getBags({ bagType: forge.pki.oids.pkcs8ShroudedKeyBag });
-
-    const certs = certBags[forge.pki.oids.certBag]?.filter((b) => b.cert) || [];
-    const keys = keyBags[forge.pki.oids.pkcs8ShroudedKeyBag]?.filter((b) => b.key) || [];
-
-    if (certs.length === 0) throw new Error('No certificates found in P12 file');
-    if (keys.length === 0) throw new Error('No private keys found in P12 file');
-
-    let signingCert = certs[0].cert!;
-
-    for (const certBag of certs) {
-      for (const keyBag of keys) {
-        const certLocalKeyId = certBag.attributes?.localKeyId?.[0];
-        const keyLocalKeyId = keyBag.attributes?.localKeyId?.[0];
-        if (certLocalKeyId && keyLocalKeyId && certLocalKeyId === keyLocalKeyId) {
-          signingCert = certBag.cert!;
-          break;
-        }
-      }
-    }
+    const { certificate } = parseP12(buffer, password);
 
     return {
-      serialNumber: BigInt('0x' + signingCert.serialNumber).toString(),
-      subject: formatRfc4514(signingCert.subject.attributes),
-      issuer: formatRfc4514(signingCert.issuer.attributes),
-      validFrom: signingCert.validity.notBefore,
-      validTo: signingCert.validity.notAfter,
+      serialNumber: BigInt('0x' + certificate.serialNumber).toString(),
+      subject: formatRfc4514(certificate.subject.attributes),
+      issuer: formatRfc4514(certificate.issuer.attributes),
+      validFrom: certificate.validity.notBefore,
+      validTo: certificate.validity.notAfter,
     };
   }
 }
